turning the variables to double solves the problem of the final result

+++ Lab_1/Question_2/src/Main.java
@@ -4,20 +4,20 @@ public class Main {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
         System.out.println("what is the value of A?");
-        int a = input.nextInt();
+        double a = input.nextInt();
         System.out.println("what is the value of B?");
-        int b = input.nextInt();
+        double b = input.nextInt();
         System.out.println("what is the value of C?");
-        int c= input.nextInt();
-        int x = quadraticFunction(a,b,c);
+        double c= input.nextInt();
+        double x = quadraticFunction(a,b,c);
 
         System.out.println("the value of X = "+x);
     }
 
-    public static int quadraticFunction(int a,int b, int c){
+    public static double quadraticFunction(double a,double b, double c){
 
-        int delta= (int) (Math.pow(b,2)- 4*a*c);
-        int x = (int) ((-(b+Math.sqrt(delta)))/2*a)/100;
+        double delta= (Math.pow(b,2)- 4*a*c);
+        double x =  ((-(b+Math.sqrt(delta)))/2*a)/100;
 
         return x;
     }
